make hearts heal player

+++ Assets/Scripts/Collectables/Collectable.cs
@@ -6,11 +6,12 @@ namespace Collectables
     // Parent class.
     public abstract class Collectable : MonoBehaviour
     {
-        public float value;
+        public int value;
         private bool isInteractable;
+
         [Tooltip("Timer before the item is interactable (by the player)")]
         public float pickupTimer = 0;
-        
+
         protected Player.Player player;
 
         protected abstract void OnCollectablePickup();
@@ -31,14 +32,14 @@ namespace Collectables
             //Disable interactable
             isInteractable = false;
             GetComponent<CircleCollider2D>().enabled = isInteractable;
-            
+
             yield return new WaitForSeconds(pickupTimer);
-            
+
             //Re-enable Interactable
             isInteractable = true;
             GetComponent<CircleCollider2D>().enabled = isInteractable;
         }
-        
+
         private void OnTriggerEnter2D(Collider2D col)
         {
             if (isInteractable)
